requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
docxtpl==0.20.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jose==1.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pillow==10.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tqdm==4.60.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
